pyproject.toml
[project]
name = "docling-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
        "docling>=2.55.0",
    "langchain-docling>=0.1.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "langchain-chroma>=0.1.0",
    "streamlit>=1.28.0",
    "streamlit-extras>=0.7.0",
    "python-dotenv>=1.0.0",
    "chromadb>0.5.5,<0.6.0",
    "tiktoken>=0.5.0",
    "pandas>=2.0.0",
    "numpy<2",
]
